dressifye_saas/requirements/base.txt
django==5.2
djangorestframework
django-allauth
PyJWT
cryptography
django-environ
django-storages[s3]
boto3
celery
redis
django-celery-beat
Pillow
requests
sentry-sdk[django]
psycopg2-binary
iyzipay
whitenoise
gunicorn
dj-database-url
django-cryptography-django5
httpx
